+++ project_code/src/erosion/flood_fill.hpp
@@ -4,6 +4,6 @@
 
 namespace floodFill {
 
-	cgp::grid_2D<int> getfloodBool(cgp::mesh const& m,int initialIndex);
+	cgp::grid_2D<short> getfloodBool(cgp::mesh const& m,int initialIndex);
 
 }
